content/canvas/test/test_size.attributes.html

<!DOCTYPE HTML>
<title>Canvas test: size.attributes</title>
<!-- Testing: width/height DOM attributes and content attributes -->
<script src="/MochiKit/packed.js"></script>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css">
<body>
<canvas id="c" width="120" height="60"><p class="fallback">FAIL (fallback content)</p></canvas>
<script>

SimpleTest.waitForExplicitFinish();
MochiKit.DOM.addLoadEvent(function () {

var canvas = document.getElementById('c');
var ctx = canvas.getContext('2d');

ok(canvas.width == 120, "canvas.width == 120");
ok(canvas.height == 60, "canvas.height == 60");
ok(canvas.getAttribute('width') == 120, "canvas.getAttribute('width') == 120");
ok(canvas.getAttribute('height') == 60, "canvas.getAttribute('height') == 60");

SimpleTest.finish();

});
</script>
